# openapi.json
{
  "components": {
...
    "schemas": {
...
      "ErrorResponse": {
        "properties": {
          "error": {
            "properties": {
              "code": {
                "type": "string"
              },
              "detail": {
                "nullable": true
              },
              "message": {
                "type": "string"
              }
            },
            "required": [
              "code",
              "message"
            ],
            "type": "object"
          }
        },
        "required": [
          "error"
        ],
        "type": "object"
      },
...
      "Manifest": {
        "properties": {
          "columns": {
            "default": [
              "id",
              "row_digest",
              "kind",
              "source",
              "payload",
              "meta",
              "signals"
            ],
            "items": {
              "type": "string"
            },
            "type": "array"
          },
          "created_at": {
            "type": "string"
          },
          "hash_algo": {
            "default": "blake3",
            "type": "string"
          },
          "kinds": {
            "additionalProperties": {
              "minimum": 0,
              "type": "integer"
            },
            "type": "object"
          },
          "name": {
            "default": null,
            "nullable": true,
            "type": "string"
          },
          "num_rows": {
            "minimum": 0,
            "type": "integer"
          },
          "schema_version": {
            "default": "1",
            "type": "string"
          },
          "version": {
            "type": "string"
          }
        },
        "required": [
          "version",
          "num_rows",
          "kinds",
          "created_at"
        ],
        "type": "object"
      },
...
      "Term": {
        "properties": {
          "aliases": {
            "default": [],
            "items": {
              "type": "string"
            },
            "type": "array"
          },
          "canonical": {
            "type": "string"
          },
          "meta": {
            "additionalProperties": {
              "nullable": true
            },
            "default": {},
            "type": "object"
          }
        },
        "required": [
          "canonical"
        ],
        "type": "object"
      },
...
      "ValidateResponse": {
        "properties": {
          "dataset": {
            "$ref": "#/components/schemas/Manifest"
          },
          "summary": {
            "$ref": "#/components/schemas/ValidateSummary"
          }
        },
        "required": [
          "summary",
          "dataset"
        ],
        "type": "object"
      },
      "ValidateSummary": {
        "properties": {
          "checked": {
            "minimum": 0,
            "type": "integer"
          },
          "invalid": {
            "minimum": 0,
            "type": "integer"
          },
          "offending_values": {
            "additionalProperties": {
              "minimum": 0,
              "type": "integer"
            },
            "default": {},
            "type": "object"
          }
        },
        "required": [
          "checked",
          "invalid"
        ],
        "type": "object"
      },
...
      "VocabulariesPage": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/VocabularyInfo"
            },
            "type": "array"
          },
          "limit": {
            "minimum": 0,
            "type": "integer"
          },
          "offset": {
            "minimum": 0,
            "type": "integer"
          },
          "total": {
            "minimum": 0,
            "type": "integer"
          }
        },
        "required": [
          "total",
          "limit",
          "offset",
          "items"
        ],
        "type": "object"
      },
      "Vocabulary": {
        "properties": {
          "dimension": {
            "type": "string"
          },
          "id": {
            "type": "string"
          },
          "meta": {
            "additionalProperties": {
              "nullable": true
            },
            "default": {},
            "type": "object"
          },
          "name": {
            "default": null,
            "nullable": true,
            "type": "string"
          },
          "source": {
            "default": null,
            "nullable": true,
            "type": "string"
          },
          "status": {
            "default": "curated",
            "enum": [
              "draft",
              "curated"
            ],
            "type": "string"
          },
          "terms": {
            "default": [],
            "items": {
              "$ref": "#/components/schemas/Term"
            },
            "type": "array"
          }
        },
        "required": [
          "dimension",
          "id"
        ],
        "type": "object"
      },
      "VocabularyInfo": {
        "properties": {
          "dimension": {
            "type": "string"
          },
          "id": {
            "type": "string"
          },
          "name": {
            "default": null,
            "nullable": true,
            "type": "string"
          },
          "num_terms": {
            "minimum": 0,
            "type": "integer"
          },
          "status": {
            "default": null,
            "enum": [
              "draft",
              "curated",
              null
            ],
            "nullable": true,
            "type": "string"
          }
        },
        "required": [
          "id",
          "dimension",
          "num_terms"
        ],
        "type": "object"
      },
      "VocabularyInput": {
        "properties": {
          "dimension": {
            "type": "string"
          },
          "meta": {
            "additionalProperties": {
              "nullable": true
            },
            "default": {},
            "type": "object"
          },
          "name": {
            "default": null,
            "nullable": true,
            "type": "string"
          },
          "source": {
            "default": null,
            "nullable": true,
            "type": "string"
          },
          "status": {
            "default": "curated",
            "enum": [
              "draft",
              "curated"
            ],
            "type": "string"
          },
          "terms": {
            "default": [],
            "items": {
              "$ref": "#/components/schemas/Term"
            },
            "type": "array"
          }
        },
        "required": [
          "dimension"
        ],
        "type": "object"
      }
    }
  },
  "info": {
    "description": "Post-training dataset control plane API.",
    "title": "databench service",
    "version": "0.0.0"
  },
  "openapi": "3.0.3",
  "paths": {
...
    "/v1/vocabularies": {
      "get": {
        "parameters": [
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 20,
              "maximum": 500,
              "minimum": 1,
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "offset",
            "required": false,
            "schema": {
              "default": 0,
              "minimum": 0,
              "nullable": true,
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/VocabulariesPage"
                }
              }
            },
            "description": "Paginated vocabularies"
          },
          "default": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Error response"
          }
        }
      }
    },
    "/v1/vocabularies/{name}": {
      "get": {
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Vocabulary"
                }
              }
            },
            "description": "Vocabulary"
          },
          "default": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Error response"
          }
        }
      },
      "put": {
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/VocabularyInput"
              }
            }
          }
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Vocabulary"
                }
              }
            },
            "description": "Vocabulary"
          },
          "default": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Error response"
          }
        }
      }
    },
    "/v1/vocabularies/{name}:derive": {
      "post": {
        "parameters": [
          {
            "in": "query",
            "name": "dataset",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "dimension",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Vocabulary"
                }
              }
            },
            "description": "Derived vocabulary"
          },
          "default": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Error response"
          }
        }
      }
    },
    "/v1/vocabularies/{name}:normalize": {
      "post": {
        "parameters": [
          {
            "in": "query",
            "name": "dataset",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "ref",
            "required": false,
            "schema": {
              "nullable": true,
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Manifest"
                }
              }
            },
            "description": "Normalized dataset manifest"
          },
          "default": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Error response"
          }
        }
      }
    },
    "/v1/vocabularies/{name}:validate": {
      "post": {
        "parameters": [
          {
            "in": "query",
            "name": "dataset",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "ref",
            "required": false,
            "schema": {
              "nullable": true,
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ValidateResponse"
                }
              }
            },
            "description": "Validation result and annotated dataset"
          },
          "default": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorResponse"
                }
              }
            },
            "description": "Error response"
          }
        }
      }
    },
...
  }
}